Replace usage of con level in combat formulas with level difference * Con level doesn't appear to be used directly on live. This change mostly helps low level players, and when fighting low orange / low red / low purple targets. It prevents big jumps in effectiveness, miss rate, and resist rate from one target to another when the level difference is only 1. It also prevents asymmetric scaling, since for example a level 51 NPC is orange to a level 50 player, but that player is still yellow to that NPC. * Critical shot's effectiveness is now lowered by 0.075 per level difference instead of 0.3 per con difference. Cap is unchanged. * Miss rate, when a NPC is involved, is now modified by a flat 1.33% per level difference instead of a flat 5% per con difference. * Miss/resist rate for new archery is now modified by a flat 1.33% per level difference instead of a flat 5% per con difference. New archery is very outdated and needs to be checked. * Miss rate is no longer reduced by a flat 0.1% per spec level (not accurate for any patch). * Base miss rate of NPCs was reduced from 25% to 18%. * Miss rate returned by `GetMissChance` now has decimals. * Base resist rate for spells is now 12.5% instead of 12% (value from Uthgard, based on live tests). * Resist rate for spells, when a NPC is involved, no longer scales off con difference. It isn’t replaced by anything since the level difference between the spell level and the target level is already used, and an extra penalty or bonus based on caster level is unconfirmed to be accurate. This increases damage against high level targets due to `AdjustDamageForHitChance` only triggering when the resist rate is at least 46%, which requires a level difference (between spell and target) of at least 67. * Resist rates returned by `CalculateToHitChance` and `CalculateSpellResistChance` now have decimals. * Triple Wield damage no longer scales off con difference. * Interrupt chance is now reduced by a flat 3% per level difference instead of a flat 15% per con difference. Still 100% against a target of the same level. * Made `CharmSpellHandler.ApplyEffectOnTarget` call `CalculateToHitChance` instead of using a base miss rate of 15% for non-pulsing spells. * Removed `PVE_SPELL_CONHITPERCENT` property. * Removed usage of `GetConLevel` for interrupt chance in `Archery.CasterIsAttacked`. New archery is very outdated and needs to be checked. * Removed usage of `GetConLevel` for block chance calculation in `ArrowOnTargetAction.Tick`. New archery is very outdated and needs to be checked. * Removed obsolete unit tests.

+++ GameServer/spells/Artifacts/CommonArtifactSpells.cs
@@ -1,22 +1,3 @@
-/*
- * DAWN OF LIGHT - The first free open source DAoC server emulator
- * 
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * 
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
- *
- */
-
 using DOL.AI.Brain;
 using DOL.GS.Effects;
 
@@ -28,10 +9,11 @@ namespace DOL.GS.Spells
     [SpellHandlerAttribute("AllStatsDebuff")]
     public class AllStatsDebuff : SpellHandler
     {
-		public override int CalculateSpellResistChance(GameLiving target)
+		public override double CalculateSpellResistChance(GameLiving target)
 		{
 			return 0;
 		}
+
         public override void OnEffectStart(GameSpellEffect effect)
         {    
      		base.OnEffectStart(effect);            
@@ -110,10 +92,11 @@ namespace DOL.GS.Spells
     [SpellHandlerAttribute("LoreDebuff")]
     public class LoreDebuff : SpellHandler
     {
- 		public override int CalculateSpellResistChance(GameLiving target)
+ 		public override double CalculateSpellResistChance(GameLiving target)
 		{
 			return 0;
 		}
+
         public override void OnEffectStart(GameSpellEffect effect)
         {
         	base.OnEffectStart(effect);      
@@ -178,11 +161,12 @@ namespace DOL.GS.Spells
     /// </summary>
     [SpellHandlerAttribute("StrengthConstitutionDrain")]
     public class StrengthConstitutionDrain : StrengthConDebuff
-    {  	
-		public override int CalculateSpellResistChance(GameLiving target)
+    {
+		public override double CalculateSpellResistChance(GameLiving target)
 		{
 			return 0;
 		}
+
         public override void OnEffectStart(GameSpellEffect effect)
         {
         	base.OnEffectStart(effect);
