+++ GameServer/spells/SiegeDirectDamageSpellHandler.cs
@@ -12,12 +12,12 @@ namespace DOL.GS.Spells
         /// </summary>
         /// <param name="target">the target of the spell</param>
         /// <returns>chance that spell will be resisted for specific target</returns>
-        public override int CalculateSpellResistChance(GameLiving target)
+        public override double CalculateSpellResistChance(GameLiving target)
         {
             return 0;
         }
 
-        public override int CalculateToHitChance(GameLiving target)
+        public override double CalculateToHitChance(GameLiving target)
         {
             return 100;
         }
